+++ database/migrations/2019_11_22_173012_create_column_city_id_in_customers.php
@@ -9,7 +9,7 @@ class CreateColumnCityIdInCustomers extends Migration
     public function up()
     {
         Schema::table('customers', function (Blueprint $table) {
-            $table->unsignedBigInteger('city_id')->after('phone');
+            $table->unsignedInteger('city_id')->after('phone');
             $table->foreign('city_id')
                 ->references('id')->on('cities')
                 ->onDelete('cascade')
